functions/requirements.txt
# functions/requirements.txt

# Web Framework
Flask==3.0.3

# Production WSGI Server (Required for Cloud Functions)
gunicorn==22.0.0

# Firebase Admin SDK for Python
firebase-admin==6.5.0

# Google AI SDK for Gemini API
google-generativeai==0.7.0

# For loading .env file (useful for local development)
python-dotenv==1.0.1

# Firebase Functions SDK (for connecting to the HTTP trigger)
firebase-functions==0.2.0  # <-- THIS IS THE MISSING PIECE